lambdas/requirements.txt
psycopg2-binary>=2.9.9
rapidfuzz>=3.6.0
boto3>=1.34.0
